Refactor ctags into separate package.
It does not scale to have ctags indexing for all languages in a single
file. The commit abstracts over the commonalities between scala and java
indexing so that we can reuse most code between the two.
